+++ app/assets/js/src/data/dayTasks/persistence/saveDayTasks.ts
@@ -1,12 +1,15 @@
 import { type PersistApi } from 'persist';
+import { saveRegister } from 'utils';
 
+import { StorageKey } from 'data/shared';
 import { dayTasksRegister } from '../dayTasksRegister';
 
 /**
  * Save the current day tasks data in memory into persistent storage.
  */
-export async function saveDayTasks(persist: PersistApi): Promise<void> {
-	const dayTasksInfo = Array.from(dayTasksRegister.entries());
-
-	await persist.set('day-tasks', dayTasksInfo);
+export function saveDayTasks(persist: PersistApi): Promise<void> {
+	return saveRegister(dayTasksRegister, {
+		persist,
+		key: StorageKey.DAY_TASKS,
+	});
 }