+++ app/assets/js/src/data/days/persistence/saveDays.ts
@@ -1,12 +1,15 @@
 import { type PersistApi } from 'persist';
+import { saveRegister } from 'utils';
 
+import { StorageKey } from 'data/shared';
 import { daysRegister } from '../daysRegister';
 
 /**
  * Save the current days data in memory into persistent storage.
  */
-export async function saveDays(persist: PersistApi): Promise<void> {
-	const daysInfo = Array.from(daysRegister.entries());
-
-	await persist.set('days', daysInfo);
+export function saveDays(persist: PersistApi): Promise<void> {
+	return saveRegister(daysRegister, {
+		persist,
+		key: StorageKey.DAYS,
+	});
 }