+++ app/assets/js/src/data/tasks/persistence/loadTasks.ts
@@ -1,5 +1,9 @@
-import { type PersistApi } from 'persist';
+import { z } from 'zod';
 
+import { type PersistApi } from 'persist';
+import { isZodSchemaType, loadRegister } from 'utils';
+
+import { StorageKey } from 'data/shared';
 import { tasksRegister } from '../tasksRegister';
 import { updateOldTaskInfo } from '../updateOldTaskInfo';
 
@@ -14,34 +18,23 @@ export async function loadTasks(
 	persist: PersistApi,
 	serialisedTasksInfo?: string
 ): Promise<void> {
-	const persistedTasksInfo = await (() => {
-		if (typeof serialisedTasksInfo !== 'undefined') {
-			return JSON.parse(serialisedTasksInfo);
+	const dataSource = serialisedTasksInfo
+		? {
+			data: serialisedTasksInfo,
 		}
+		: {
+			persist,
+			key: StorageKey.TASKS,
+		};
 
-		return persist.get('tasks');
-	})();
-
-	if (typeof persistedTasksInfo === 'undefined') {
-		tasksRegister.clear();
-		return;
-	}
-
-	if (!(
-		Array.isArray(persistedTasksInfo) &&
-		persistedTasksInfo.every((el): el is [number, unknown] => (
-			Array.isArray(el) &&
-			el.length === 2 &&
-			typeof el[0] === 'number'
-		))
-	)) {
-		throw new Error(`Persisted tasks data is invalid: ${serialisedTasksInfo}`);
-	}
-
-	const newTasksInfo = persistedTasksInfo.map(
-		([taskId, taskInfo]) => [taskId, updateOldTaskInfo(taskInfo)] as const
+	const isValidTaskEntry = isZodSchemaType(
+		z.tuple([z.number(), z.unknown()])
 	);
 
-	tasksRegister.clear();
-	tasksRegister.set(newTasksInfo);
+	return loadRegister(
+		tasksRegister,
+		dataSource,
+		isValidTaskEntry,
+		updateOldTaskInfo,
+	);
 }