+++ app/assets/js/src/data/templates/persistence/saveTemplates.ts
@@ -1,12 +1,15 @@
 import { type PersistApi } from 'persist';
+import { saveRegister } from 'utils';
 
+import { StorageKey } from 'data/shared';
 import { templatesRegister } from '../templatesRegister';
 
 /**
  * Save the current templates data in memory into persistent storage.
  */
 export async function saveTemplates(persist: PersistApi): Promise<void> {
-	const templatesInfo = Array.from(templatesRegister.entries());
-
-	return persist.set('templates', templatesInfo);
+	return saveRegister(templatesRegister, {
+		persist,
+		key: StorageKey.TEMPLATES,
+	});
 }