+++ app/assets/js/src/persist/idb.ts
@@ -1,58 +1,8 @@
 import type { PersistApi } from 'persist/PersistApi';
-
-const dbName = 'orange-twist';
-const objectStoreName = 'orange-twist';
-
-let db: IDBDatabase | null = null;
-/**
- * Get a handle to the database, opening it if it wasn't
- * already open.
- */
-function getDatabase(): Promise<IDBDatabase> {
-	return new Promise((resolve, reject) => {
-		if (db) {
-			resolve(db);
-			return;
-		}
-
-		const request = indexedDB.open(dbName);
-
-		// Create object store when database is first created
-		request.addEventListener('upgradeneeded', () => {
-			request.result.createObjectStore(objectStoreName);
-		});
-
-		// Handle success
-		request.addEventListener('success', () => {
-			db = request.result;
-			resolve(request.result);
-		});
-
-		// Handle errors
-		request.addEventListener('error', reject);
-		request.addEventListener('blocked', reject);
-	});
-}
-
-/**
- * Handle the repetitive paperwork like starting a transaction
- * and wrapping things in a Promise.
- */
-function doTransaction(
-	mode: IDBTransactionMode,
-	callback: (objectStore: IDBObjectStore) => IDBRequest
-): Promise<unknown> {
-	return new Promise((resolve, reject) => {
-		getDatabase().then((db) => {
-			const transaction = db.transaction(objectStoreName, mode);
-			const objectStore = transaction.objectStore(objectStoreName);
-			const request = callback(objectStore);
-
-			request.addEventListener('success', () => resolve(request.result));
-			request.addEventListener('error', () => reject(request.error));
-		}).catch(reject);
-	});
-}
+import {
+	ObjectStoreName,
+	doDatabaseTransaction,
+} from 'utils';
 
 /**
  * A {@linkcode PersistApi} interface for working with the
@@ -60,22 +10,25 @@ function doTransaction(
  */
 export const idb: PersistApi = {
 	async set(key, data) {
-		await doTransaction(
+		await doDatabaseTransaction(
 			'readwrite',
+			ObjectStoreName.DATA,
 			(objectStore) => objectStore.put(data, key)
 		);
 	},
 
 	get(key) {
-		return doTransaction(
+		return doDatabaseTransaction(
 			'readonly',
+			ObjectStoreName.DATA,
 			(objectStore) => objectStore.get(key)
 		);
 	},
 
 	async delete(key) {
-		await doTransaction(
+		await doDatabaseTransaction(
 			'readwrite',
+			ObjectStoreName.DATA,
 			(objectStore) => objectStore.delete(key)
 		);
 	},